requirements-dev.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
pydantic-settings
psycopg[binary]
alembic
PyJWT==2.10.1
Ruff
python-multipart
boto3
pillow
redis
rq
